requirements.txt
customtkinter
pillow
exifread
imagehash
tqdm
pandas
matplotlib